refactor(autoware_scenario_selector): merge default_scenario_selector and extra_scenario_selector to scenario_gate
