--- planning/autoware_default_scenario_selector/launch/dummy_default_scenario_selector_lane_driving.launch.xml
@@ -1,16 +1,0 @@
-<launch>
-  <arg name="input_lane_driving_trajectory"/>
-  <arg name="output_scenario"/>
-  <arg name="output_trajectory"/>
-
-  <arg name="input_parking_trajectory" default=""/>
-  <arg name="input_lanelet_map" default=""/>
-  <arg name="input_route" default=""/>
-  <arg name="input_odometry" default=""/>
-  <arg name="is_parking_completed" default=""/>
-
-  <node pkg="topic_tools" exec="relay" name="scenario_trajectory_relay" output="log" args="$(var input_lane_driving_trajectory) $(var output_trajectory)"/>
-
-  <arg name="cmd" default="ros2 topic pub $(var output_scenario) autoware_internal_planning_msgs/msg/Scenario '{current_scenario: LANEDRIVING, activating_scenarios: [LANEDRIVING]}'"/>
-  <executable cmd="$(var cmd)" name="scenario_pub" output="screen" shell="true"/>
-</launch>
