--- planning/autoware_default_scenario_selector/package.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>autoware_default_scenario_selector</name>
-  <version>0.46.0</version>
-  <description>The autoware_default_scenario_selector ROS 2 package</description>
-  <maintainer email="taiki.tanaka@tier4.jp">Taiki Tanaka</maintainer>
-  <maintainer email="tomoya.kimura@tier4.jp">Tomoya Kimura</maintainer>
-  <maintainer email="shumpei.wakabayashi@tier4.jp">Shumpei Wakabayashi</maintainer>
-  <maintainer email="fumiya.watanabe@tier4.jp">Fumiya Watanabe</maintainer>
-  <maintainer email="takamasa.horibe@tier4.jp">Takamasa Horibe</maintainer>
-  <maintainer email="satoshi.ota@tier4.jp">Satoshi Ota</maintainer>
-  <maintainer email="takayuki.murooka@tier4.jp">Takayuki Murooka</maintainer>
-  <maintainer email="wei@turing-drive.com">ChiaWei Li</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake_auto</buildtool_depend>
-  <buildtool_depend>autoware_cmake</buildtool_depend>
-
-  <depend>autoware_lanelet2_extension</depend>
-  <depend>autoware_map_msgs</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_planning_test_manager</depend>
-  <depend>autoware_route_handler</depend>
-  <depend>autoware_utils</depend>
-  <depend>nav_msgs</depend>
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
-  <depend>tf2</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>tf2_ros</depend>
-  <depend>tier4_planning_msgs</depend>
-  <depend>autoware_scenario_gate</depend>
-
-  <exec_depend>ros2cli</exec_depend>
-  <exec_depend>topic_tools</exec_depend>
-
-  <test_depend>ament_cmake_ros</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>autoware_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
