--- planning/autoware_extra_scenario_selector/CMakeLists.txt
@@ -1,32 +1,0 @@
-cmake_minimum_required(VERSION 3.14)
-project(autoware_extra_scenario_selector)
-
-find_package(autoware_cmake REQUIRED)
-autoware_package()
-
-find_package(pluginlib REQUIRED)
-
-pluginlib_export_plugin_description_file(autoware_scenario_gate plugins.xml)
-
-ament_auto_add_library(${PROJECT_NAME} SHARED
-  src/node.cpp
-)
-
-rclcpp_components_register_node(${PROJECT_NAME}
-  PLUGIN "autoware::scenario_selector::ExtraScenarioSelectorN"
-  EXECUTABLE ${PROJECT_NAME}_node
-)
-
-if(BUILD_TESTING)
-  ament_add_ros_isolated_gtest(test_${PROJECT_NAME}
-    test/test_${PROJECT_NAME}_node_interface.cpp
-  )
-  target_link_libraries(test_${PROJECT_NAME}
-    ${PROJECT_NAME}
-  )
-endif()
-
-ament_auto_package(INSTALL_TO_SHARE
-  launch
-  config
-)
