--- planning/autoware_extra_scenario_selector/config/extra_scenario_selector.param.yaml
@@ -1,8 +1,0 @@
-/**:
-    ros__parameters:
-      update_rate: 10.0
-      th_max_message_delay_sec: 1.0
-      th_arrived_distance_m: 1.0
-      th_stopped_time_sec: 1.0
-      th_stopped_velocity_mps: 0.01
-      enable_mode_switching: true
