--- planning/autoware_extra_scenario_selector/launch/extra_scenario_selector.launch.xml
@@ -1,33 +1,0 @@
-<launch>
-  <arg name="input_lane_driving_trajectory" default="/planning/scenario_planning/lane_driving/trajectory"/>
-  <arg name="input_parking_trajectory" default="/planning/scenario_planning/parking/trajectory"/>
-  <arg name="input_waypoint_following_trajectory" default = "/planning/scenario_planning/waypoint_following/trajectory"/>
-  <arg name="input_lanelet_map" default="/map/vector_map"/>
-  <arg name="input_route" default="/planning/mission_planning/route"/>
-  <arg name="input_odometry" default="/localization/kinematic_state"/>
-  <arg name="input_operation_mode_state" default="/system/operation_mode/state"/>
-  <arg name="output_scenario" default="/planning/scenario_selector/scenario"/>
-  <arg name="output_trajectory" default="/planning/scenario_selector/trajectory"/>
-  <arg name="is_parking_completed" default="/planning/scenario_planning/parking/is_completed"/>
-  <arg name="is_waypoint_following_completed" default="/planning/scenario_planning/waypoint_following/is_completed"/>
-
-  <!-- Parameter -->
-  <arg name="config_file" default="$(find-pkg-share autoware_extra_scenario_selector)/config/extra_scenario_selector.param.yaml"/>
-
-  <node pkg="autoware_extra_scenario_selector" exec="autoware_extra_scenario_selector_node" name="extra_scenario_selector" output="screen">
-    <remap from="input/lane_driving/trajectory" to="$(var input_lane_driving_trajectory)"/>
-    <remap from="input/parking/trajectory" to="$(var input_parking_trajectory)"/>
-    <remap from="input/waypoint_following/trajectory" to="$(var input_waypoint_following_trajectory)"/>
-    <remap from="input/lanelet_map" to="$(var input_lanelet_map)"/>
-    <remap from="input/route" to="$(var input_route)"/>
-    <remap from="input/odometry" to="$(var input_odometry)"/>
-    <remap from="input/operation_mode_state" to="$(var input_operation_mode_state)"/>
-    <remap from="is_parking_completed" to="$(var is_parking_completed)"/>
-    <remap from="is_waypoint_following_completed" to="$(var is_waypoint_following_completed)"/>
-
-    <remap from="output/scenario" to="$(var output_scenario)"/>
-    <remap from="output/trajectory" to="$(var output_trajectory)"/>
-    <!-- Parameter -->
-    <param from="$(var config_file)"/>
-  </node>
-</launch>
