+++ planning/autoware_scenario_gate/package.xml
@@ -16,6 +16,15 @@
   <depend>autoware_planning_msgs</depend>
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_utils</depend>
+  <depend>autoware_lanelet2_extension</depend>
+  <depend>autoware_map_msgs</depend>
+  <depend>autoware_route_handler</depend>
+  <depend>nav_msgs</depend>
+  <depend>tf2</depend>
+  <depend>tf2_geometry_msgs</depend>
+  <depend>tf2_ros</depend>
+  <depend>tier4_planning_msgs</depend>
+  <depend>autoware_adapi_v1_msgs</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
