--- planning/autoware_scenario_selector/launch/dummy_scenario_selector_waypoint_following.launch copy.xml
@@ -1,16 +1,0 @@
-<launch>
-  <arg name="input_waypoint_following_trajectory"/>
-  <arg name="output_scenario"/>
-  <arg name="output_trajectory"/>
-
-  <arg name="input_lane_driving_trajectory" default=""/>
-  <arg name="input_lanelet_map" default=""/>
-  <arg name="input_route" default=""/>
-  <arg name="input_odometry" default=""/>
-  <arg name="is_waypoint_following_completed" default=""/>
-
-  <node pkg="topic_tools" exec="relay" name="scenario_trajectory_relay" output="log" args="$(var input_waypoint_following_trajectory) $(var output_trajectory)"/>
-
-  <arg name="cmd" default="ros2 topic pub $(var output_scenario) tier4_planning_msgs/msg/Scenario '{current_scenario: Waypoint_following, activating_scenarios: [Waypoint_following]}'"/>
-  <executable cmd="$(var cmd)" name="scenario_pub" output="screen" shell="true"/>
-</launch>
